Fix up version number, since 0.7.2-OPOWER had already been released.
Updated Environment.java to print out the endpoint paths correctly when
the JAX-RS annotations are on the interface instead of a concrete class.

Updated JDBI version to 2.51 to get bugfixes, a major one with trying to
open a connection when an onDemand DAO was being finalized.  ISP-596

Tested the build under Java 7 as well, and it works.  Placed properties
in order to migrate to Java 7 when the time comes.